requirements.txt
streamlit
openai
google-generativeai
python-dotenv
pypdf
python-docx
gTTS
